feat: combines mobile and desktop footers
